move UpgradesPage SSG to server dir
